Sorting out `Breed`/`BreedManager` nonsense
This was my one remaining bit of unfinished business in the initial shot
of the engine rewrite (which I accidentally forgot about).  Turns out
that I _can_ separate `Breed` from `BreedManager`, despite my snarky
commentary that made it sound a lot tougher than it actually was.  Truth
be told, it _was_ annoying to reconcile that circular dependency.  The
engine is better off for this change, though--a million fold.

This change got a bunch of tests passing, but there are still plenty
failing.