Fixing when patch variable updates are triggered
This was a major source of errors across test suites.  I think the
majority of tests that were still failing were failing because of this.
Overall, it's because IntelliJ's Vim emulator doesn't know how to preserve
indentation across copying and pasting, even though the proper "copy"
and "paste" functionalities do it correctly.  Thanks, plugin author!