Marking "CoffeeScript ranges" off of my list
Yeah... turns out that tossing out CoffeeScript ranges isn't as
desirable as I had thought.  I mean, I know that `..` and `...` are
cryptic and confused (and the wrong one was actually used in the code
that I'm replacing here), but `_(0).range(derp).forEach((x) ->` has its
own readability problems in many places.  It's a bit unfortunate, but I
think the `for` loops are superior for readability in many of the cases,
sucky as they are.