Moving bootstrapping code to the right place
The old code was functioning, but the extra three dependencies on
`Workspace` shouldn't have been there.  The reason that they got put
there at all was a bit of forgetfulness on my part, I guess.  I doesn't
help that the way RequireJS handles this stuff is a bit confusing to me.
For some reason, certainly modules can't be loaded at certain times if
they weren't a part of the initial burst of module-loading.  Seems
pretty silly to me, but whatevs....