Expanding usages of `notImplemented`
It now returns a function for all of the primitive's arguments to
pointlessly be passed into.  At first, I weighed whether or not it was
worthwhile to be writing out arguments that are ignored.  On the one
hand, they are basically useless and wasting processor cycles.  On the
other hand, they provide the necessary arguments for plotting (etc.)
that someone who wants to hand-implement the relevant functionality will
want.  I initially did it without allowing the arguments, but it caused
an "RNG out of sync" error, so the arguments now get written out and
evaluated, then ignored.  Whatevs.