Making sure not to return a value from `_.forEach` calls
This bit me when merging in Seth's giant pull request that got a bunch
more models working.

'Ethnocentrism' was failing, because certain turtles weren't getting
their shapes set correctly.  I was able to fairly quickly pinpoint the
commit and the function (`_makeTurtleCopy`) where this was occurring,
and it seemed that `arr.forEach(f)` would yield the correct result, but
`_(arr).forEach(f)` wouldn't, for some reason unbeknownst to me.

As it turns out, after **quite a bit** of investigation and
befuddlement (followed by a fearful glance at the documentation),
Lodash's `forEach` aborts early if one of the calls of `f` evaluates to
`false`.  `turtle.setVariable(n, <x: boolean>)` was returning a boolean,
and `x` was sometimes `false`, causing `setVariable` to return `false`
(since _it_ doesn't properly end with `return`), causing `f` to return
`false`, causing the loop to break early and not copy the turtle
variable that determined which shape a turtle got in the next tick.
Fun, fun....