Making turtles take IDs in their constructor
Seems more sane and sensical and FP to me.  I don't like the idea of
clients creating turtles, stuffing them into the world, and expecting
the world to insert an ID later.  Instead, they pass the world a
function for creating the turtle, and that function will be called with
and ID when the world creates one.  Magical!