Letting tick updates go through `World` again
The situation was quickly appearing to get out of hand.  It's nice to be
able to put isolated functionality/variables/updates into small
subobjects, but it was going to be irritating to have to special-case
`Updater` every time we wanted to add another isolated box of
functionality.  With this, we have a pattern for how to effectively
avoid that, I think.