Simplifying ways to get/set an agent's variables
Sadly, I don't see a good way that I can get rid of
`[gs]etPatchVariable`, since turtles can (for whatever reason) be asked
about patch variables (which requires routing that variable-related call
through the patch that the turtle is on).  I wish I could get rid of
that, since it would make the variable-getting/-setting interface
totally unified and (fairly) simple.  But the new situation is much
better than the old one, and it's not apparent to me that improving it
further would be worth my time at the moment.