Making my `VariableManager`s use `Object.defineProperty`
All that silly code that I wrote was redundant with this
natively-implemented thing.  This way is faster, and requires
maintaining less code, at the cost of being blatantly mutable (see
new `refineBy`).  It's appears to be a change for the best, though.
Thanks goes to Bryan for letting me know that this exists.