Shuffling around some functions to do away with `Prims#AgentOrSet`
Union types suck, and the code that utilized this union type was ugly.
I think that it's _a bit_ better now, but I _definitely_ think that
things are now better suited for future improvements on this front.

Admittedly, I replaced the code for `turtlesOn` with something that uses
only a slightly less-narrow union type, but... I like this version more,
and it's not apparent to me that there's a good way to implement
`turtlesOn` as a method of agentsets; the fact that it returns a fresh
turtleset is problematic.  It's doable, but in a pretty ugly way.