Switching from RequireJS to Browserify
I've abandoned RequireJS and I could hardly be happier.  RequireJS has
made me miserable nearly every time I've tried to use it, and, when this
overhauled version of the engine wouldn't work in Galapagos, that was
the last straw for me.  Browserify is _so_ much less magical and,
consequently, way easier to deal with.  Long live Browserify!

In this moment, I am euphoric.  Not because of any phony jrburke's
blessing, but because I am enlightened by my own Browserification.

Model dumps updated in next commit