Fixing bug with new Browserify system
The bug was that you needed to run Grunt twice in order for the aliases
to be properly formed.  This is because Grunt apparently evaluates all
of a task's values once the pipeline starts, rather than lazily.  So our
alias list would be essentially empty, then `coffee` would run, put all
of the files in place that we wanted to glob over, and then the
near-empty alias list would be applied.  Then, the second time Grunt was
run, the compiled CoffeeScript files would be detected when the pipeline
kicked off, aliases would be generated from those, and then the
CoffeeScript would be compiled and the aliases would be used for
Browserify.

Now, we make sure to update the alias list appropriately.