Fixing SpiderMonkey and V8 to work again
The Browserify changes broke them.  The problem is that, for a reason
that I don't fully understand, Lodash, in this Browserified build,
insists on having access to the global object, and it expects that
object to be called "window".  Therefore, all of the engines need to
have `window` set properly in them.