Ensuring that `clear-all` calls `clear-patches` (fixes #107)
This was actually a pretty neat bug.  Looks like, back in f8533aa, the
intended code in the implementation of `clear-all` was `clearPatches`,
but `createPatches` was typed instead, and it worked fine, given the
state of the engine at the time.  However, when I implemented patch
neighbor caching in fa1fa80, the crouching tiger beetle/hidden dragonfly
leapt from the shadows and struck!  This is because the `Topology` was
keeping a cache of patches, but, after `world.createPatches`, the cached
patches were no longer in the set of patches being tracked by the
`World`.  Subtle.

Oh, also in this commit is a dependency on a new version of Headless,
which contains a test that failed before this commit and now passes.