Having turtles only poll their own links
Amusingly, turtles were tracking their own links, but doing pretty much
nothing with that value.

On the more serious side of things, this change should offer a
substantial performance benefit for models that use links, and it was
really a necessary change for testability, since iterating over
`world.links()` would ultimately mess with RNG state and the iteration
order of the resultant agentsets.