Fixing bug in agents dying while their agentset is being processed
This addresses a case where some agent, A, would ask some agent, B, to
die while iterating over an agentset that contains both A and B.
Previously, B would still run the command block, even while dead.  Now,
when a shufflerator detects an item that does not meet its criteria for
validity, that item is ignored.  This behavior is compliant with JVM
NetLogo's behavior.