Attempting to supplant Lodash with some handmade collections
The motivation for this is that the latest version bump on Lodash
caused BZ Benchmark to run 25% slower.  I wanted to see how much better
I could write my own collections that did only and exactly what I
wanted.

This is the result.  There are a few bugs left unresolved in this
commit, and the code is about 18% slower than the old Lodash code
(making it a tad faster than the current Lodash version).  I don't
think that I speedup that I got is worth the cost of maintaining
our own collections library, so I'm going to set this aside for now.