Infrastructure: Updating JSON lib versions
Thank Bokonon that someone finally took over the Grunt Browserify
project and fixed it up, enabling me to, at long last, update this
thing for the first time in 9 fricken months!

Also, as of this commit, Coffeelint is now explicitly included as a
dependency, because NPM is telling me that I'll need to do so in NPM 3
(whenever that's even supposed to come out).  Previously, t was being
pulled in by grunt-coffeelint.