Refactor: Substantial memory usage reduction in agents
It seems that it's unwise to have each agent possess its own copy of
every getter/setter function.  Instead, put them in a centralized place,
and call out to them.  Leads to pretty huge memory usage reductions
for large numbers of agents.

Also, this allowed me to move a lot of the noisy crap out of
`turtle.coffee`--something that had been bothering me for quite a while.
I won't say that it's excellent now, but I think I like this better.