Test/Refactor: Remove dumb test; test elsewhere
The test that I removed is something that I think is pretty pointless.
It doesn't really test Tortoise.  It is testing headless to make sure
that its JSON is matching some static JSON (which doesn't necessarily
equal the JSON coming out of Tortoise).  As a result, I find that that
test has negative value.  So it's dead now.