Test/Refactor: Remove docking garbage that was fixed many JVMs ago
This is actually pretty major.  Not only does it speed up the docking
tests and remove some complexity, but it actually fixes a critical bug.
This code was dumping out numeric things in slots named, e.g. "_0"
instead of "0".  This led to `AgentModel` not doing what it's supposed
to, and essentially dropping and not comparing updates in the `world`
and `observer`.  This is because `AgentModel` was looking for things in
slot "0".  So... good riddance.