Test/Bugfix: Fix up regex
I thought that `^` would match the beginning of the line,
but it seems to only be matching the beginning of the
string, in this case.  Most models that use extensions
list them on the first line, so it's not a huge problem,
but some don't, and this fixes the regex for those.