Infrastructure/Tests: Auto-update extensions instead of deleting
The models library will also be using the extension manager to load
extensions in Jenkins now, so we can no longer wipe the extensions
folder when Tortoise does its testing.  Auto-updating still isn't as
ideal as specifying the extension versions to use exactly, but we can be
reasonably sure we want to use the latest version of any extension in
the library for the time being.