Feature/Parity: Add runtime type checking layer, basic math checks
These changes are based on Jason's work on the `wip-middle-layer`
branch.  Enough had changed since that work was done that I wanted to
start from scratch.  All the good ideas are his, and the bad ideas
mine.

The checking layer should have all input and result checking, and all
the I18N calls should live in this layer, rather than being sprinkled
throughout the engine.  The checks should align closely to the
primitives, and delegate the work to the appropriate engine components.

This also adds a start of an I18N effort.

As part of the refactor we also had to add a couple checks for the list
prims, but those are not yet complete.

There was a `validateNumber()` check for the `approximate-rgb` prim
method `ColorModel.nearestColorNumberOfRGB()`, but that was determined
to not be necessary, so it's removed without replacement.  Noted here
for posterity if something comes up with it in the future.