Refactor: Change the type checker to better fit runtime error checks
For runtime error type checking messages we need a way to easily specify
the types that a primtiive expects to receive as arguments as well as a
way to generate a nice description of the actual type of a value when it
does not match.

I also switched the checker to not generate throw-away objects as part of
normal quick checks like `isNumber()` and `isString()`.