Bugfix/Test: Deal with bytes in [0, 255], in 'encode' tests
Previously, the bytes in the tests were within [-128, 127], and that was fine, and had parity with desktop's implementation of the extension.  However, the real JS APIs actually produce values in [0, 255].  The Java-based APIs (from desktop, and from our polyfill here) prefer the other range, since Java only has signed bytes.

In order to get these two systems talking the same language, I changed the tests here, the polyfills here, and the implementation in desktop, to convert the values to [0, 255].