Test/Refactor: Remove unnecessary freebies
It looks like these tests got moved into the 'reporters' category in
NetLogo/NetLogo, at some point, and then maybe they got fixed, here in
this repo, when we updated the Headless dependency in this repository?

Whatever the case, these lines don't do anything, and the tests with
these names in Headless run fine in this repository now.